requirements.txt
annotated-types==0.7.0
cffi==2.0.0
numpy==2.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
SoundCard==0.4.5
soundfile==0.13.1
typing-inspection==0.4.2
typing_extensions==4.15.0
